requirement.txt
numpy
opencv-python
scikit-image
matplotlib
Pillow